refactor: extract FAISS file ingest helpers
